+++ packages/terra-clinical-data-grid/src/terra-dev-site/test/clinical-data-grid/CustomHeightDataGrid.test.jsx
@@ -86,6 +86,7 @@ class CustomHeightDataGrid extends React.Component {
     };
   }
 
+  /* eslint-disable react/forbid-dom-props */
   render() {
     return (
       <div id="custom-height-data-grid" style={{ height: '100%', width: '100%' }}>
@@ -104,6 +105,7 @@ class CustomHeightDataGrid extends React.Component {
       </div>
     );
   }
+  /* eslint-enable react/forbid-dom-props */
 }
 
 export default CustomHeightDataGrid;